send On branch circle_2 Your branch is up to date with origin/circle_2. 	modified:   circle_2/_push_swap/Makefile 	deleted:    circle_2/_push_swap/checker_linux 	modified:   circle_2/_push_swap/libft/ft_atoi.c 	modified:   circle_2/_push_swap/libstack/stack_utils.c 	modified:   circle_2/_push_swap/push_swap.c 	modified:   circle_2/_push_swap/push_swap.h 	modified:   circle_2/_push_swap/sorting_algo/sort.c 	modified:   circle_2/_push_swap/stack_op/stack_op.c 	modified:   circle_2/_push_swap/utils/start.c 	circle_0/libft/ 	circle_1/get_next_line/ 	circle_2/.DS_Store 	circle_2/_push_swap/.DS_Store 	circle_2/_push_swap/checker_Mac 	circle_2/_push_swap/ft_printf/ft_on_flag_action.o 	circle_2/_push_swap/ft_printf/ft_parse_flags.o 	circle_2/_push_swap/ft_printf/ft_printf.o 	circle_2/_push_swap/ft_printf/ft_putchar_fd.o 	circle_2/_push_swap/ft_printf/ft_putnbr_base.o 	circle_2/_push_swap/ft_printf/ft_putstr_fd.o 	circle_2/_push_swap/ft_printf/ft_strlen.o 	circle_2/_push_swap/ft_printf/libftprintf.a 	circle_2/_push_swap/libft/ft_atoi.o 	circle_2/_push_swap/libft/ft_bzero.o 	circle_2/_push_swap/libft/ft_calloc.o 	circle_2/_push_swap/libft/ft_count_words.o 	circle_2/_push_swap/libft/ft_free_tab.o 	circle_2/_push_swap/libft/ft_intlen.o 	circle_2/_push_swap/libft/ft_isalnum.o 	circle_2/_push_swap/libft/ft_isalpha.o 	circle_2/_push_swap/libft/ft_isascii.o 	circle_2/_push_swap/libft/ft_isdigit.o 	circle_2/_push_swap/libft/ft_islower.o 	circle_2/_push_swap/libft/ft_isprint.o 	circle_2/_push_swap/libft/ft_isupper.o 	circle_2/_push_swap/libft/ft_itoa.o 	circle_2/_push_swap/libft/ft_memchr.o 	circle_2/_push_swap/libft/ft_memcmp.o 	circle_2/_push_swap/libft/ft_memcpy.o 	circle_2/_push_swap/libft/ft_memmove.o 	circle_2/_push_swap/libft/ft_memset.o 	circle_2/_push_swap/libft/ft_putchar_fd.o 	circle_2/_push_swap/libft/ft_putendl_fd.o 	circle_2/_push_swap/libft/ft_putnbr_fd.o 	circle_2/_push_swap/libft/ft_putstr_fd.o 	circle_2/_push_swap/libft/ft_split.o 	circle_2/_push_swap/libft/ft_strchr.o 	circle_2/_push_swap/libft/ft_strdup.o 	circle_2/_push_swap/libft/ft_striteri.o 	circle_2/_push_swap/libft/ft_strjoin.o 	circle_2/_push_swap/libft/ft_strlcat.o 	circle_2/_push_swap/libft/ft_strlcpy.o 	circle_2/_push_swap/libft/ft_strlen.o 	circle_2/_push_swap/libft/ft_strmapi.o 	circle_2/_push_swap/libft/ft_strncmp.o 	circle_2/_push_swap/libft/ft_strnstr.o 	circle_2/_push_swap/libft/ft_strrchr.o 	circle_2/_push_swap/libft/ft_strtrim.o 	circle_2/_push_swap/libft/ft_substr.o 	circle_2/_push_swap/libft/ft_tolower.o 	circle_2/_push_swap/libft/ft_toupper.o 	circle_2/_push_swap/libft/libft.a 	circle_2/_push_swap/libstack/stack_utils.o 	circle_2/_push_swap/push-swap-tester-42/ 	circle_2/_push_swap/push_swap 	circle_2/_push_swap/push_swap.o 	circle_2/_push_swap/sorting_algo/is_sorted.o 	circle_2/_push_swap/sorting_algo/radix.o 	circle_2/_push_swap/sorting_algo/sort.o 	circle_2/_push_swap/stack_op/stack_op.o 	circle_2/_push_swap/utils/error.o 	circle_2/_push_swap/utils/start.o to github

+++ circle_2/_push_swap/libft/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By:  mchenava < mchenava@student.42lyon.fr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/10 16:43:55 by mchenava          #+#    #+#             */
-/*   Updated: 2022/12/12 14:09:45 by  mchenava        ###   ########.fr       */
+/*   Updated: 2023/01/09 12:13:22 by  mchenava        ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,5 +35,5 @@ long	ft_atoi(const char *nptr)
 	}
 	if (nptr[i] != '\0')
 		return (LONG_MAX);
-	return ((int)(result * sign));
+	return ((result * sign));
 }
